pyproject.toml
[tool.poetry]
name = "razrabot-todo-list"
version = "0.1.0"
description = ""
authors = ["BogdanBarylo <barylob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
mysql-connector = "^2.2.9"
mysql-connector-python = "^8.4.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
